refactor:feature 구조 변경
