+++ src/features/auth/ui/index.ts
@@ -3,3 +3,5 @@
 // export * from './SignupForm';
 // export * from './LogoutButton';
 export { default as AuthLayout } from './AuthLayout';
+export { default as TextField } from './TextField';
+export { default as SubmitButton } from './SubmitButton';
