--- src/features/note/api/index.ts
@@ -1,3 +1,0 @@
-// Note API 세그먼트 Public API
-// export * from './note-api';
-// export * from './queries';
