--- src/features/note/model/index.ts
@@ -1,4 +1,0 @@
-export * from './queryKeys';
-export * from './types';
-// export * from './note-store';
-// export * from './note-selectors';
