--- src/features/note/ui/index.ts
@@ -1,4 +1,0 @@
-// Note UI 세그먼트 Public API
-// export * from './CreateNoteForm';
-// export * from './EditNoteForm';
-// export * from './NoteActions';
