--- src/features/profile/api/index.ts
@@ -1,3 +1,0 @@
-// Profile API 세그먼트 Public API
-// export * from './profile-api';
-// export * from './queries';
