--- src/features/profile/model/index.ts
@@ -1,4 +1,0 @@
-export * from './queryKeys';
-export * from './types';
-// export * from './profile-store';
-// export * from './profile-selectors';
