--- src/features/profile/model/queryKeys.ts
@@ -1,4 +1,0 @@
-export const profileQueryKeys = {
-  all: ['profile'] as const,
-  current: () => [...profileQueryKeys.all, 'current'] as const,
-} as const;
