--- src/features/profile/ui/index.ts
@@ -1,4 +1,0 @@
-// Profile UI 세그먼트 Public API
-// export * from './EditProfileForm';
-// export * from './ProfileSettings';
-// export * from './ChangePasswordForm';
