--- src/features/side-bar/api/index.ts
@@ -1,3 +1,0 @@
-// Side-bar API 세그먼트 Public API
-// export * from './sidebar-api';
-// export * from './queries';
