--- src/features/side-bar/index.ts
@@ -1,4 +1,0 @@
-// Side-bar 피처 Public API
-export * from './model';
-// export * from './api';
-// export * from './ui';
