--- src/features/side-bar/model/index.ts
@@ -1,4 +1,0 @@
-export * from './queryKeys';
-export * from './types';
-// export * from './sidebar-store';
-// export * from './sidebar-selectors';
