--- src/features/side-bar/ui/index.ts
@@ -1,4 +1,0 @@
-// Side-bar UI 세그먼트 Public API
-// export * from './SideBar';
-// export * from './Navigation';
-// export * from './MenuItems';
