--- src/features/todo/api/index.ts
@@ -1,3 +1,0 @@
-// Todo API 세그먼트 Public API
-// export * from './todo-api';
-// export * from './queries';
