--- src/features/todo/index.ts
@@ -1,4 +1,0 @@
-// Todo 피처 Public API
-export * from './model';
-// export * from './api';
-// export * from './ui';
