--- src/features/todo/model/index.ts
@@ -1,4 +1,0 @@
-export * from './queryKeys';
-export * from './types';
-// export * from './todo-store';
-// export * from './todo-selectors';
