+++ src/shared/ui/index.ts
@@ -1,1 +1,1 @@
-export { default as TextField } from './TextField';
+export { default as TextField } from '../../features/auth/ui/TextField';
